rearci-iac/requirements.txt
aws-cdk-lib==2.214.0
constructs>=10.0.0,<11.0.0
boto3
requests
beautifulsoup4
